# bookreader-openapi.yaml
openapi: "3.0.3"
info:
  title: "bookreader API"
  description: "bookreader API"
  version: "1.0.0"
servers:
  - url: "https://bookreader"
paths:
...
  /api/book/{googleBookId}:
    get:
      summary: "GET api/book/{googleBookId}"
      operationId: "findBookByGoogleBookID"
      parameters:
        - name: "googleBookId"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
  /api/book/{id}:
    delete:
      summary: "DELETE api/book/{id}"
      operationId: "deleteBook"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
...
  /api/bookshelves/recommended/books:
    put:
      summary: "PUT api/bookshelves/recommended/books"
      operationId: "addRecommendedBook"
      parameters:
        - name: "userID"
          in: "query"
          required: true
          schema:
            type: "number"
            format: "int64"
        - name: "recommenderID"
          in: "query"
          required: true
          schema:
            type: "number"
            format: "int64"
        - name: "bookID"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
...
  /api/bookshelves/{name}/books:
    put:
      summary: "PUT api/bookshelves/{name}/books"
      operationId: "addBook"
      parameters:
        - name: "name"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "bookID"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "userID"
          in: "query"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
...
  /api/bookshelves/{name}/remove:
    put:
      summary: "PUT api/bookshelves/{name}/remove"
      operationId: "deleteBook"
      parameters:
        - name: "name"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "userID"
          in: "query"
          required: true
          schema:
            type: "number"
            format: "int64"
        - name: "bookID"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
...
  /api/personalTimeline:
    get:
      summary: "GET api/personalTimeline"
      operationId: "getPersonalTimeline"
      parameters:
        - name: "userid"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
...
  /api/publicTimeline:
    get:
      summary: "GET api/publicTimeline"
      operationId: "getPublicTimeline"
      parameters:
        - name: "userids"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
...
  /api/reply:
    get:
      summary: "GET api/reply"
      operationId: "userReplyComment"
      parameters:
        - name: "eventid"
          in: "query"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
    post:
      summary: "POST api/reply"
      operationId: "userReplyComment"
      responses:
        "200":
          description: "OK"
...
  /api/review/book/{googleBookId}:
    get:
      summary: "GET api/review/book/{googleBookId}"
      operationId: "findReviewsByGoogleBookID"
      parameters:
        - name: "googleBookId"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
  /api/review/user/{userId}:
    get:
      summary: "GET api/review/user/{userId}"
      operationId: "findReviewsByUserID"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
  /api/review/{id}:
    delete:
      summary: "DELETE api/review/{id}"
      operationId: "deleteReview"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
...